final with minor bugs to be fixed

+++ public/src/components/Logout.jsx
@@ -1,20 +1,26 @@
-import styled from "styled-components";
+import React from "react";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
 import { BiPowerOff } from "react-icons/bi";
-
-
+import styled from "styled-components";
+import axios from "axios";
+import { logoutRoute } from "../utilities/APIRoutes";
 export default function Logout() {
-    const navigate = useNavigate();
-    const handleClick = async () => {
-        localStorage.clear();
-        navigate('/login');
+  const navigate = useNavigate();
+  const handleClick = async () => {
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
     }
+  };
   return (
     <Button onClick={handleClick}>
-        <BiPowerOff />
+      <BiPowerOff />
     </Button>
-  )
+  );
 }
 
 const Button = styled.button`
